start implementing emulator tests

+++ src/lib/pis.c
@@ -92,7 +92,7 @@ u64 pis_operand_size_max_unsigned_value(pis_operand_size_t operand_size) {
     if (bits == 64) {
         return UINT64_MAX;
     } else {
-        return (1 << bits) - 1;
+        return ((u64) 1 << bits) - 1;
     }
 }
